refactor: 도서 검색 및 상세 데이터 레이어 패키지 구조 개선
- Naver, Aladin 등 특정 플랫폼 이름이 포함된 클래스 및 패키지명을 일반적인 명칭으로 변경
- 도서 상세 정보 관련 데이터 소스를 별도 패키지로 분리
- API 응답 모델 구조 단순화 및 재정의
