+++ app/src/main/java/com/boostcamp/and03/data/datasource/remote/book_detail/BookDetailRemoteDataSource.kt
@@ -1,0 +1,8 @@
+package com.boostcamp.and03.data.datasource.remote.book_detail
+
+import com.boostcamp.and03.data.model.response.BookLookUpResponse
+
+interface BookDetailRemoteDataSource {
+
+    suspend fun loadBookPage(itemId: String): BookLookUpResponse
+}
