+++ app/src/main/java/com/boostcamp/and03/data/di/datasource/BookSearchDataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.boostcamp.and03.data.di.datasource
 
-import com.boostcamp.and03.data.datasource.remote.search.naver.BookSearchRemoteDataSource
-import com.boostcamp.and03.data.datasource.remote.search.naver.BookSearchRemoteDataSourceImpl
+import com.boostcamp.and03.data.datasource.remote.search.BookSearchRemoteDataSource
+import com.boostcamp.and03.data.datasource.remote.search.BookSearchRemoteDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
