--- app/src/main/java/com/boostcamp/and03/data/model/response/AladinBookInfo.kt
@@ -1,9 +1,0 @@
-package com.boostcamp.and03.data.model.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AladinBookInfo(
-    @SerialName("itemPage") val itemPage: Int
-)
